LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/traverseTo-cross-document-expected.txt



FAIL cross-document traverseTo(), back(), and forward() assert_equals: expected 1 but got 0
